adapter: remove expensive confirm_leadership call
This commit removes a potentially expensive `confirm_leadership` call on
the sequencing thread. The call needs to potentially block for a long
time, as it waits for the persist catalog to catch up and then applies
all the observed changes. During that time, the environment would be
unresponsive to commands, so it seems like a good idea to get rid of the
call.