adapter: take read hold on target during replacement MV sequencing
When sequencing the creation of a replacement MV, we have to make sure
it's as-of is greater than the read frontier of the output shard (i.e.
the target MV's storage collection).

Previously, we did this just by bumping the selected as-of by the
target's read frontier, assuming that we already had a read hold on the
target. However, before doing MV timestamp selection, we only have read
holds on the replacement MV's inputs. Which made it possible for the
target MV's read frontier to be downgraded before we managed to install
the replacement MV's dependency on it in `StorageCollections` tripping a
sanity check there.

The fix is to properly take a read hold on the target MV's storage
collection. We also pull the special handling for replacement MVs out of
`select_timestamp`, so we have both acquiring of the read hold and
bumping of the `storage_as_of` in the same place.