adapter,storage: fix creation of storage collections for replacement MVs
This commit fixes a bug in how the storage collections for replacement
MVs are created. Previously, the sequencer would set the as-of of these
collections to the target MV's `since` + 1. This would cause
`StorageCollections::create_collections_for_bootstrap` to open a
`SinceHandle` with that as-of and advance the shard's critical since to
that as-of as well. At the same time, the target MV's collection would
still have its original `since`, now less than the persist shard's
since, which could cause "cannot server requested as-of" panics when
trying to render a dataflow reading from that MV.

The fix chosen here is to make `StorageCollections` check whether the
new collection is a non-primary one, and if so, avoid downgrading its
critical since. We already do that elsewhere, so I think not doing it
during collection creation was just an oversight.

Furthermore, this commit removes the special timestamp selection logic
for replacement MVs, that advances that as-of to target `since` + 1.
That logic isn't needed and was only added due to a misunderstanding of
the storage controller's requirements around dependent collections.