storage: skip wrong assert for primary dependencies
This commit fixes a bug where creating a replacement MV could trigger an
assert in the storage controller that's meant to check an invariant
about the frontiers of sources and their remap shards. Replacement MVs
don't require that same invariant.

I don't really like this solution, it is a quick fix that reduces code
quality. There are some ideas for proper fixes, all involving getting
rid of the "primary" concept, but they require more discussion and
bigger refactors.