catalog-protos: remove most uses of the `Empty` type
In the protobuf-derived format the `Empty` type was necessary because
protobuf cannot represent union enum variants, so we'd end up with,
e.g., `GlobalId::Explain(Empty {})`. This can now just be
`GlobalId::Explain` and the `Empty` type isn't needed anymore.

Note that we can't fully delete it because the JSON representation of
the audit log objects cannot change, so we need to keep `Empty` around
for those.