caro build: Silence s3tables warning
Seen locally during `cargo build --all`:

> warning: Patch `iceberg-catalog-s3tables v0.7.0 (https://github.com/MaterializeInc/iceberg-rust.git?rev=cf1e6e0c9de3#cf1e6e0c)` was not used in the crate graph.
Check that the patched package version and available features are compatible
with the dependency requirements. If the patch has a different version from
what is locked in the Cargo.lock file, run `cargo update` to use the new
version. This may also occur with an optional dependency that is not enabled.