sql,catalog: revert `replacing` column from SHOW MVs
This commits reverts the addition of the `replacing` column from the
output of SHOW MATERIALIZED VIEWS. This column is fed by a new builtin
table, and our 0dt upgrade logic currently cannot handle changing an
existing builtin dataflow starting to depend on a new builtin table. It
gets stuck in the caught-up check instead.