catalog: index most of `pg_authid`
pgbouncer by default queries `pg_authid` to authenticate users. This
query is slow because it reads from unindexed sources and computes an
expensive dataflow. Ideally we would fix this by indexing it, but that's
not possible because two columns are computed using unmaterializable
functions. So instead we introduce a new `pg_authid_core` view that
excludes these columns and can be indexed. `pg_authid` is then defined
on top of this view.