catalog: the mother of all migrations
This commit adds a big migration from the previous protobuf-structured
format to the new denoised format.

Some new special handling needs to be introduced in the `objects!` macro
in `upgrade.rs`, due to the new format having less indirection than the
old format. We'll be able to remove that again eventually, when we can
delete the old object definitions.