storage: use port 0 for DNS resolution
We were using a `DUMMY_DNS_PORT` of 11111 to pass to address override
functions that expect a `SocketAddr`. The assumption was that the port
is ignored, so it might as well be recognizable.

However, recent changes in `hyper-util` make it so that the port is now
used to override the default port, if no explicit port is specified in
the URL. Thus, using port 11111 is wrong now and we should use port 0
instead to defer to the protocol's default port.

This commit changes all our usages of `DUMMY_DNS_PORT` to use port 0
instead. This gives us the previous behavior. No attempts are made to
make use of the new feature to actually override the port (e.g. when
setting up an SSH tunnel).