[persist] Fetch fewer diffs during GC (#33267)
While we recently switched to limiting the amount of seqnos that can be
GC'd in one pass, we'd still end up fetching all the diffs in every GC
run, because we didn't have a good way to get both the initial state and
a set of follow-up diffs otherwise. This PR switches to grabbing the
appropriate rollup from the current state... which we're already loading
earlier in the method.

### Motivation

It's twice now we've seen high memory pressure due to a stuck shard
suddenly getting unstuck. This helps us bound the amount of memory used
in such circumstances...

### Tips for reviewer

A few refactorings as well, which hopefully speak for themselves.

I put the new behaviour behind the active-compaction flag; that flag is
currently CI-only and pairs well with the changes here I think.